+++ src/ct/Lexer.h
@@ -48,9 +48,9 @@ namespace CodeTools
 struct LineInfo
 {
    const size_t begin;  // offset where line starts; it ends at a CRLF
-   int depth;           // lexical level for indentation
+   int8_t depth;        // lexical level for indentation
    bool ctorBraceInit;  // set for brace initialization of member
-   bool continuation;   // set if code continues from the previous line
+   bool continuation;   // incremented if code continues from the previous line
    bool mergeable;      // set if code can merge with another line
    bool c_comment;      // set if the line lies within a C-style comment
    LineType type;       // line's type
@@ -658,7 +658,7 @@ private:
    //  first line in the range.  If MERGE is false, the lines are marked as
    //  not mergeable.
    //
-   void SetDepth(int depth1, int depth2, bool merge = true);
+   void SetDepth(int8_t depth1, int8_t depth2, bool merge = true);
 
    //  Returns true if the left brace at curr_ is a brace initialization.
    //  PREV is the character that preceded the left brace.
